backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
pgvector==0.2.4
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
feedparser==6.0.11
httpx==0.26.0
openai==1.10.0
apscheduler==3.10.4
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
beautifulsoup4==4.12.3
aiofiles==23.2.1
python-slugify==8.0.1
authlib==1.3.0
itsdangerous==2.1.2
